+++ modules/python_engine/builtin/cpp/__pyenv__Builtin.cpp
@@ -7,30 +7,37 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "handle_propertiesBuiltin.hpp"
+#include "__pyenv__Builtin.hpp"
 #include "Error.hpp"
 #include "i18n.hpp"
-#include "PredefinedErrorMessages.hpp"
+#include "nlsBuildConfig.h"
 #include "InputOutputArgumentsCheckers.hpp"
+#include "PythonEnvironment.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
 ArrayOfVector
-Nelson::HandleGateway::handle_propertiesBuiltin(int nLhs, const ArrayOfVector& argIn)
+Nelson::Python_engineGateway::__pyenv__Builtin(int nLhs, const ArrayOfVector& argIn)
 {
-    ArrayOfVector retval;
-    nargincheck(argIn, 1, 1);
-    ArrayOf param1 = argIn[0];
-    if (param1.isHandle()) {
-        auto* obj = (HandleGenericObject*)param1.getContentAsHandleScalar();
-        if (obj) {
-            retval << ArrayOf::toCellArrayOfCharacterColumnVectors(obj->getProperties());
-        } else {
-            Error(_W("Invalid handle."));
-        }
+    nargincheck(argIn, 0, 4);
+    nargoutcheck(nLhs, 0, 1);
+
+    PythonEnvironment* pythonEnvironment = PythonEnvironment::getInstance();
+
+    if (argIn.size() == 0) {
+
+    } else if (argIn.size() == 4) {
+        pythonEnvironment->setVersion(argIn[0].getContentAsWideString());
+        pythonEnvironment->setExecutable(argIn[1].getContentAsWideString());
+        pythonEnvironment->setLibrary(argIn[2].getContentAsWideString());
+        pythonEnvironment->setHome(argIn[3].getContentAsWideString());
+        pythonEnvironment->saveCurrentState();
     } else {
-        Error(ERROR_WRONG_ARGUMENT_1_TYPE_FUNCTION_HANDLE_EXPECTED);
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
     }
+
+    ArrayOfVector retval;
+    retval << ArrayOf::handleConstructor((HandleGenericObject*)pythonEnvironment);
     return retval;
 }
 //=============================================================================