+++ modules/types/src/include/Types.hpp
@@ -153,6 +153,7 @@ using wstringVector = std::vector<std::wstring>;
 #define NLS_HANDLE_AFTEREACHFUTURE_CATEGORY_STR "AfterEachFuture"
 #define NLS_HANDLE_FEVALQUEUE_CATEGORY_STR "FevalQueue"
 #define NLS_HANDLE_QOBJECT_CATEGORY_STR "QObject"
+#define NLS_HANDLE_PYTHON_ENVIRONMENT_CATEGORY_STR "PythonEnvironment"
 //=============================================================================
 } // namespace Nelson
 //=============================================================================